added CRUD for room

+++ src/controller/taskController.js
@@ -1,11 +1,12 @@
 const TaskModel = require("../model/taskModel");
-const mongoose = require("mongoose");
 
 const getAllTasks = async (req, res) => {
   try {
     const Tasks = await TaskModel.find({});
 
-    return res.status(200).send({ status: true, message: Tasks });
+    return res
+      .status(200)
+      .send({ status: true, count: Tasks.length, message: Tasks });
   } catch (error) {
     return res.status(500).send({ status: false, message: error.message });
   }
@@ -16,8 +17,10 @@ const getRandomNumberedTasks = async (length) => {
     let data = [];
     const tasks = await TaskModel.find({});
     if (tasks)
-      for (let index = 0; index < length; index++)
-        data.push(tasks[Math.floor(Math.random() * tasks.length)]);
+      for (let index = 0; index < length; index++) {
+        const val = tasks[Math.floor(Math.random() * tasks.length)];
+        if (!data.includes(val)) data.push(val);
+      }
 
     return data;
   } catch (error) {
@@ -60,6 +63,12 @@ const getTaskById = async (req, res) => {
 const insertTask = async (req, res) => {
   try {
     const data = req.body;
+    const oldTask = await TaskModel.find({ taskName: data.taskName });
+    if (oldTask)
+      return res.status(400).send({
+        status: false,
+        message: "Task with same name already exists.",
+      });
 
     const Task = await TaskModel.create(data);
 
@@ -71,13 +80,44 @@ const insertTask = async (req, res) => {
   }
 };
 
+const insertTaskBulk = async (req, res) => {
+  try {
+    let errTasks = [],
+      successTasks = [];
+
+    const body = req.body;
+    if (!body)
+      return res
+        .status(400)
+        .send({ status: false, message: "Please provide valid body params." });
+
+    await body.forEach(async (item, index) => {
+      const oldTask = await TaskModel.find({ taskName: item.taskName });
+      if (oldTask) errTasks.push(item);
+      else {
+        const Task = await TaskModel.create(item);
+        if (Task) successTasks.push(item);
+        else errTasks.push(item);
+      }
+    });
+
+    return res.status(200).send({
+      status: true,
+      message: "Tasks added successfully.",
+      errorTasks: errTasks,
+    });
+  } catch (error) {
+    return res.status(500).send({ status: false, message: error.message });
+  }
+};
+
 const updateTask = async (req, res) => {
   try {
     const { id } = req.params;
-    // if (!validator.isValidObjectId(id))
-    return res
-      .status(400)
-      .send({ status: false, message: "Please provide valid ID" });
+    if (!id)
+      return res
+        .status(400)
+        .send({ status: false, message: "Please provide valid ID" });
 
     const data = req.body;
 
@@ -94,10 +134,10 @@ const updateTask = async (req, res) => {
 const deleteTask = async (req, res) => {
   try {
     const { id } = req.params;
-    // if (!validator.isValidObjectId(id))
-    return res
-      .status(400)
-      .send({ status: false, message: "Please provide valid ID" });
+    if (!id)
+      return res
+        .status(400)
+        .send({ status: false, message: "Please provide valid ID" });
 
     const Task = await TaskModel.findByIdAndDelete(id);
 
@@ -115,6 +155,7 @@ module.exports = {
   getTasksByUserId,
   getTaskById,
   insertTask,
+  insertTaskBulk,
   updateTask,
   deleteTask,
 };
